# contracts/education/openapi.yaml
openapi: 3.0.3
info:
  title: Education Service API
  version: 0.1.0
  description: |
    学员服务 BFF/业务 API 契约（对应 specs/education-service-agent/design.md §5）。
    约定：JSON camelCase；金额为整数分；时间为 ISO8601 UTC；studentId 永远由服务端从会话取得，
    请求体中不出现。写操作接受 Idempotency-Key；状态更新携带 expectedRevision。
servers:
  - url: /api/v1
...
paths:
...
  /conversations/{conversationId}/messages:
    get:
      tags: [chat]
      operationId: listMessages
      summary: 断线后据此恢复最终消息与任务状态（V1 不重放 token）；也用于排队/接管期间轮询对方消息
      parameters:
        - {$ref: '#/components/parameters/ConversationId'}
        - name: after
          in: query
          required: false
          description: |
            不传：返回最近 200 条消息（快照）。传某条消息的 id：只返回它之后的新消息（增量，同样最多 200 条），
            给轮询用，避免排队/接管期间每次都重新拉整个窗口。这个 id 查不到（游标过期、传了别的会话的消息 id 等）
            时退化成快照，不报错。
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: 消息与运行状态
          content:
            application/json:
              schema:
                type: object
                required: [items, run, pendingConfirmation, mode, handoff]
                properties:
                  items:
                    description: 不传 after 时是最近 200 条快照；传 after 时是它之后的新消息（可能为空数组）。
                    type: array
                    items: {$ref: '#/components/schemas/Message'}
                  mode:
                    description: 会话此刻由谁应答。bot 机器人；queued 已请求人工、排队中；human 老师接管中；closed 已关闭。刷新页面后据此恢复"排队中 / 老师处理中"的界面。
                    type: string
                    enum: [bot, queued, human, closed]
                  handoff:
                    description: 进行中（排队或已接管）的接管记录；没有则为 null。已结束（released）的不在这里。
                    type: object
                    nullable: true
                    allOf: [{$ref: '#/components/schemas/Handoff'}]
                  pendingConfirmation:
                    description: |
                      这个会话里待确认的申请草稿及其确认卡（形状同 application.confirmation 事件的载荷）；没有则为 null。
                      用于断线重连后找回确认卡——它在实时事件流里发出过，但不在消息里。
                      条件：草稿仍是 draft、由这个会话里的某次运行起草（来源取自已验签的内部工作证，不由请求体或模型指定）、
                      确认卡未使用/未撤销/未过期；多张时取最新一张。
                    nullable: true
                    type: object
                    required: [applicationId, confirmationId, revision, expiresAt, summary]
                    properties:
                      applicationId: {type: string, format: uuid}
                      confirmationId: {type: string, format: uuid}
                      revision: {type: integer, minimum: 0}
                      expiresAt: {type: string, format: date-time}
                      summary: {type: object, additionalProperties: true}
                  run:
                    description: 最近一次运行；从未运行过为 null。租约（90 秒）过期的运行中 run 会在读取时被判为 failed/LEASE_EXPIRED，客户端不会永远看到"生成中"。
                    nullable: true
                    type: object
                    required: [id, kind, status, errorCode]
                    properties:
                      id: {type: string, format: uuid}
                      kind: {type: string, enum: [message, resume]}
                      status: {type: string, enum: [running, completed, failed]}
                      errorCode: {type: string, nullable: true}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      tags: [chat]
      operationId: sendMessage
      summary: 发送消息，响应为 text/event-stream（POST fetch 流，事件见 events.schema.json）
      description: |
        同一会话同一 clientMessageId 只会处理一次：原运行已完成则回放一个 message.completed（同一 messageId），
        原运行已失败则回放 run.error（要重试请换新的 clientMessageId），原运行仍在处理中返回 409 RUN_IN_PROGRESS。
        运行的生命周期不绑在这条连接上：连接断开后服务端仍会跑完并落库，之后用 GET messages 查询结果。
        Agent 不可达等"开流前"的失败返回普通错误响应（503），不是 SSE。

        会话正在排队（queued）或由老师接管（human）时：消息只被记录（老师看得到），**不启动机器人**（AC-012），
        响应仍是 SSE，但只有一个 handoff.status 事件（payload.mode 是此刻的会话状态）；该事件不属于任何 Agent 运行，
        runId 取被记录消息的 id，仅用于客户端关联。同一个 clientMessageId 重试只记录一条；接管结束之后再重试一条
        当时被记录的消息，也只返回"已记录"（handoff.status，mode 为 bot），不会补一次机器人回复。
        会话已关闭（closed）返回 409 INVALID_STATE。
      parameters: [{$ref: '#/components/parameters/ConversationId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientMessageId, text]
              properties:
                clientMessageId: {type: string, minLength: 1, maxLength: 100}
                text: {type: string, minLength: 1, maxLength: 4000}
      responses:
        '200':
          description: SSE；每个 data 行是 StreamEvent JSON
          content:
            text/event-stream:
              schema: {type: string}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409':
          description: 同会话已有生成任务（RUN_IN_PROGRESS），或会话已关闭（INVALID_STATE）
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '422': {$ref: '#/components/responses/Invalid'}
        '429': {$ref: '#/components/responses/Budget'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /conversations/{conversationId}/resume:
    post:
      tags: [chat]
      operationId: resumeConversation
      summary: 学员在界面上处理了确认卡之后，让 Agent 读取申请的最新状态并回话；响应为 text/event-stream
      description: |
        请求体为空。"确认"这件事本身已经由 POST /applications/{id}/confirm 完成，这个接口不携带、也不采信任何确认内容
        （不把"确认"文本当成授权）：Agent 恢复后会重新向业务 API 读取申请状态再回话。
        没有可恢复的确认（从未产生确认卡，或已恢复过）返回 409 INVALID_STATE；会话正在生成返回 409 RUN_IN_PROGRESS；
        会话正在排队或由老师接管返回 409 HANDOFF_ACTIVE（恢复会让机器人开口，接管期间不允许；学员在业务 API 上的确认本身不受影响）。
      parameters: [{$ref: '#/components/parameters/ConversationId'}]
      responses:
        '200':
          description: SSE；最后一个事件是 message.completed（含 messageId），助手回复已落库
          content:
            text/event-stream:
              schema: {type: string}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
  parameters:
...
    ConversationId: {name: conversationId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    Unauthorized:
      description: 未登录
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Forbidden:
      description: 角色无权限
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: 不存在，或他人私人资源（不区分）
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Invalid:
      description: 参数错误
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Conflict:
      description: 状态或版本冲突
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Budget:
      description: 预算/限流
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Unavailable:
      description: 依赖不可用
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, INVALID_STATE, REVISION_CONFLICT,
                     STALE_CONFIRMATION, IDEMPOTENCY_KEY_REUSED, RUN_IN_PROGRESS, HANDOFF_ACTIVE,
                     BUDGET_EXCEEDED, DEPENDENCY_UNAVAILABLE, INTERNAL]
            message: {type: string}
            requestId: {type: string}
            details: {type: object, additionalProperties: true}
...
    Message:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id: {type: string}
        role: {type: string, enum: [user, assistant, teacher, system]}
        content: {type: string}
        runId: {type: string, nullable: true}
        createdAt: {type: string, format: date-time}
    Handoff:
      type: object
      required: [id, conversationId, status, revision]
      properties:
        id: {type: string, format: uuid}
        conversationId: {type: string, format: uuid}
        status: {type: string, enum: [queued, claimed, released]}
        teacherId: {type: string, format: uuid, nullable: true}
        summary:
          type: string
          nullable: true
          description: |
            交接摘要，写给老师看的内部说明：由 Agent 用确定性代码从已核验的工具结果和学员原话拼出，不含猜测。
            **只在老师接口（claim / release）的响应里返回**；学员侧接口（POST handoff、GET messages）不含这个字段。
            学员直接点"转人工"、没有 Agent 参与时为 null。
        reason: {type: string, nullable: true}
        revision: {type: integer}
security:
  - sessionCookie: []